requirements.txt
faiss-cpu
sentence-transformers
numpy
tqdm
python-dotenv
rich
